Data source health check: Refactor software tests into individual files
Also split off the test fixtures into a separate file.
